feat: rename pixelpencil.js to pencil.js
Rename pixelpencil.js tool file to match simplified naming convention.

This aligns the code filename with the simplified UI name "Pencil"
instead of "Wacky Pencil".
